+++ cmd/stop.go
@@ -267,10 +267,12 @@ func stopServicesWithConfirmation(ctx context.Context, config *configs.Config, s
 	if len(servicesToStop) == 0 {
 		return nil
 	}
-	question := fmt.Sprintf("To perform this update, we need to kill these services: %v. Is this ok? Y/N", servicesToStop)
-	ok, _ := AskUserToContinue(ctx, question)
-	if !ok {
-		return fmt.Errorf("user did not accept confirmation to stop services")
+	if !config.Force {
+		question := fmt.Sprintf("To perform this update, we need to kill these services: %v. Is this ok? Y/N", servicesToStop)
+		ok, _ := AskUserToContinue(ctx, question)
+		if !ok {
+			return fmt.Errorf("user did not accept confirmation to stop services")
+		}
 	}
 	return stopServices(ctx, servicesToStop, config)
 }
